Move post-canonical pipeline to shared file
All diseases will go through this same pipeline
